refactor: modularize workforce pipeline
