+++ docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased — Blueprint Taxonomy v2
 
+- Refactored the workforce pipeline: extracted scheduling, payroll, market, intent,
+  trait, and telemetry logic into dedicated modules under `src/workforce/**`, kept
+  each file under 500 LOC, and routed all workforce side-effects through a single
+  telemetry batch emitter so the tick stage now orchestrates pure functions only.
 - Added import hygiene guardrails: `simple-git-hooks` pre-commit runs lint + import resolution specs, ESLint now blocks local
   `.js` suffixes in TypeScript, and a Vitest probe dynamically `import()`s engine schemas/pipeline stages to catch bad specifiers
   before CI.
